Arduino sketch for snake on Sparkfun RG LED matrix
We had a hack day at work years ago, and this was my team's entry. I remember winning,
most other teams just spelled out a message. I made a "button" via a wire off A0 and GND,
and touching them together would cause the snake to turn (I'm not totally sure this is
safe, but it definitely worked).

There's no dying in the game - the board wraps around and the snake easily passes through
itself. The score is kept on the bottom row of LEDs, counting up in binary.